Remove device info and use reactConfig instead
