+++ android-app/settings.gradle
@@ -1,4 +1,1 @@
 include ':xnative'
-
-include ':react-native-device-info'
-project(':react-native-device-info').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-device-info/android')
