+++ android-app/xnative/build.gradle
@@ -75,8 +75,4 @@ android {
 
 dependencies {
     api "com.facebook.react:react-native:${packageInfo.dependencies.'react-native'}"
-    api("react-repo:react-native-device-info:${packageInfo.dependencies.'react-native-device-info'}") {
-        exclude group: "com.facebook.react"
-        exclude group: "com.google.android.gms"
-    }
 }
