+++ android/app/build.gradle
@@ -152,10 +152,6 @@ repositories {
 }
 
 dependencies {
-    implementation (project(':react-native-device-info')) {
-      exclude group: "com.google.android.gms"
-      exclude group: "com.facebook.react"
-    }
     implementation "com.facebook.react:react-native:0.55.4"  // From node_modules
     implementation "com.facebook.stetho:stetho:1.5.0"
 }
