+++ android/settings.gradle
@@ -1,4 +1,2 @@
 rootProject.name = 'storybooknative'
-include ':react-native-device-info'
-project(':react-native-device-info').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-device-info/android')
 include ':app'
