--- packages/date-publication/__mocks__/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "env": {
-    "jest": true
-  }
-}
