--- packages/date-publication/__mocks__/react-native-device-info.js
@@ -1,8 +1,0 @@
-let timezone = "Europe/London";
-
-module.exports = {
-  setMockTimezone: t => {
-    timezone = t;
-  },
-  getTimezone: jest.fn(() => timezone)
-};
