+++ packages/date-publication/__tests__/ios/date-publication-non-london.ios.test.js
@@ -1,11 +1,7 @@
-import shared from "../shared-london";
-
-const realIntl = Intl;
+import shared from "../shared-non-london";
 
 const dateBST = "2017-01-01T14:32:00.000Z";
 const dateGMT = "2017-07-01T14:32:00.000Z";
 
 shared(dateBST);
 shared(dateGMT);
-
-global.Intl = realIntl;
