+++ packages/date-publication/__tests__/ios/date-publication.ios.test.js
@@ -1,14 +1,7 @@
-import rndiMock from "react-native-device-info";
-import shared from "../shared";
-
-const londonTimezone = () => {
-  rndiMock.setMockTimezone("Europe/London");
-};
-
-const nonLondonTimezone = () => {
-  rndiMock.setMockTimezone("Europe/Kiev");
-};
+import shared from "../shared-london";
 
 const dateBST = "2017-01-01T14:32:00.000Z";
+const dateGMT = "2017-07-01T14:32:00.000Z";
 
-shared(dateBST, { londonTimezone, nonLondonTimezone });
+shared(dateBST);
+shared(dateGMT);
