+++ packages/date-publication/__tests__/mock-timezone-london.js
@@ -1,0 +1,5 @@
+jest.mock("react-native", () => {
+  const rn = require.requireActual("react-native");
+  rn.NativeModules.ReactConfig = { timezone: "Europe/London" };
+  return rn;
+});
