+++ packages/date-publication/__tests__/mock-timezone-non-london.web.js
@@ -1,0 +1,5 @@
+global.Intl = {
+  DateTimeFormat: () => ({
+    resolvedOptions: () => ({ timeZone: "Europe/Kiev" })
+  })
+};
