+++ packages/date-publication/__tests__/shared-london.js
@@ -1,14 +1,14 @@
 import React from "react";
 import TestRenderer from "react-test-renderer";
 import { iterator } from "@times-components/test-utils";
+import "./mock-timezone-london";
 import DatePublication from "../src/date-publication";
 
-export default (date, { londonTimezone, nonLondonTimezone }) => {
+export default date => {
   const tests = [
     {
       name: "date and TIMES publication",
       test() {
-        londonTimezone();
         const testInstance = TestRenderer.create(
           <DatePublication date={date} publication="TIMES" />
         );
@@ -19,7 +19,6 @@ export default (date, { londonTimezone, nonLondonTimezone }) => {
     {
       name: "date and SUNDAYTIMES publication",
       test() {
-        londonTimezone();
         const testInstance = TestRenderer.create(
           <DatePublication date={date} publication="SUNDAYTIMES" />
         );
@@ -30,7 +29,6 @@ export default (date, { londonTimezone, nonLondonTimezone }) => {
     {
       name: "date and no given publication",
       test() {
-        londonTimezone();
         const testInstance = TestRenderer.create(
           <DatePublication date={date} />
         );
@@ -41,20 +39,8 @@ export default (date, { londonTimezone, nonLondonTimezone }) => {
     {
       name: "date and no day",
       test() {
-        londonTimezone();
         const testInstance = TestRenderer.create(
           <DatePublication date={date} showDay={false} />
-        );
-
-        expect(testInstance).toMatchSnapshot();
-      }
-    },
-    {
-      name: "date and TIMES publication with non-London timezone",
-      test() {
-        nonLondonTimezone();
-        const testInstance = TestRenderer.create(
-          <DatePublication date={date} publication="TIMES" />
         );
 
         expect(testInstance).toMatchSnapshot();
