+++ packages/date-publication/__tests__/shared-non-london.js
@@ -1,0 +1,22 @@
+import React from "react";
+import TestRenderer from "react-test-renderer";
+import { iterator } from "@times-components/test-utils";
+import "./mock-timezone-non-london";
+import DatePublication from "../src/date-publication";
+
+export default date => {
+  const tests = [
+    {
+      name: "date and TIMES publication with non-London timezone",
+      test() {
+        const testInstance = TestRenderer.create(
+          <DatePublication date={date} publication="TIMES" />
+        );
+
+        expect(testInstance).toMatchSnapshot();
+      }
+    }
+  ];
+
+  iterator(tests);
+};
