+++ packages/date-publication/src/date/index.js
@@ -28,12 +28,6 @@ const getUTCTime = date => {
   );
 };
 
-const isLondonTimezone = () => {
-  try {
-    return getTimezone() === "Europe/London";
-  } catch (err) {
-    return false;
-  }
-};
+const isLondonTimezone = () => getTimezone() === "Europe/London";
 
 export { isLondonTimezone, getUTCTime, isBST };
